Split direct emit gated dependency tests
