+++ docs/COMPLETION_AUDIT.md
@@ -2717,6 +2717,10 @@ and do not assume Phase 4 is ready without evidence.
   `src/typechecker/self_type_validation/expressions.rs`, preserving the same
   context diagnostics while keeping declaration task collection and type
   reference validation separate.
+- Direct `zen emit build.zen` gated dependency validation tests now live in
+  `tests/integration/cli_build/emit_direct_validation/gated_dependencies.rs`,
+  preserving the gated dependency diagnostic coverage while keeping direct emit
+  graph validation cases grouped by failure mode.
 - Effect checking, typed allocator semantics, actors in std integration,
   JSON/YAML IR boundaries, and broader build graph execution remain gated by
   `docs/V1_SPEC.md`.
